+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Replace device `metadata` with `attributes`.
+- BREAKING: Remove the 0.90.1 Astarte client API and introduce a clean, idiomatic API.
+  See [#33](https://github.com/astarte-platform/astarte-go/issues/33).
 
 ## [0.90.1] - 2021-03-03
 ### Changed